% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_kmers.R
\name{extract_kmers}
\alias{extract_kmers}
\title{Extract continuous k-mers from a sequence}
\usage{
extract_kmers(seq, n)
}
\arguments{
\item{seq}{Sequence from which k-mers will be extracted}

\item{n}{Lenght of k-mers}
}
\description{
Extracts continuous k-mers of a given length from a single sequence in a FASTA format
}
\details{
Extracts continuous k-mers of a given length from a single sequence
}
