% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_seq}
\alias{new_seq}
\title{Class for amino acid sequences}
\usage{
new_seq(x, name, ...)
}
\arguments{
\item{x}{sequence}

\item{name}{name of a sequence}

\item{...}{additional arguments}
}
\description{
Stores the sequence
}
\details{
Class for storing the sequence
}
