% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.nuc}
\alias{print.nuc}
\title{Print function for seq class}
\usage{
\method{print}{nuc}(x, ...)
}
\arguments{
\item{x}{sequence}

\item{...}{additional arguments}
}
\description{
Prints the sequence
}
\details{
Prints the sequence
}
