% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta_file.R
\name{read_fasta_file}
\alias{read_fasta_file}
\title{Read fasta file}
\usage{
read_fasta_file(file)
}
\arguments{
\item{file}{File containing sequences in a FASTA format}
}
\description{
Reads in fasta files and stores sequences
}
\details{
Function that reads fasta files and stores sequences in a named list
}
\examples{
more_seq <- read_fasta_file(system.file(package = "SeqPakiet", "more_seq.fa"))
seq <- read_fasta_file(system.file(package = "SeqPakiet", "seq.fa"))
}
