% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.kmers_list}
\alias{summary.kmers_list}
\title{Summary function for list of kmers}
\usage{
\method{summary}{kmers_list}(object, ...)
}
\arguments{
\item{object}{sequences}

\item{...}{additional arguments}
}
\description{
Prints the composition of a sequences
}
\details{
Prints the composition of sequences
}
