% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.seq}
\alias{summary.seq}
\title{Summary function for seq class}
\usage{
\method{summary}{seq}(object, ...)
}
\arguments{
\item{object}{sequence}

\item{...}{additional arguments}
}
\description{
Prints the composition of a sequence
}
\details{
Prints the composition of a sequence
}
