% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.seq_list}
\alias{summary.seq_list}
\title{Summary function for list of sequences of seq class}
\usage{
\method{summary}{seq_list}(object, ...)
}
\arguments{
\item{object}{sequences}

\item{...}{additional arguments}
}
\description{
Prints the composition of a sequences
}
\details{
Prints the composition of sequences
}
